chore: move remaining cowswap utils to cowswap-utils
