+++ packages/swapper/src/swappers/CowSwapper/filterAssetIdsBySellable/filterAssetIdsBySellable.ts
@@ -1,9 +1,9 @@
 import type { AssetId } from '@shapeshiftoss/caip'
 import type { Asset } from '@shapeshiftoss/types'
 
+import { SUPPORTED_CHAIN_IDS } from '../../../cowswap-utils/constants'
 import { isNativeEvmAsset } from '../../utils/helpers/helpers'
 import { COWSWAP_UNSUPPORTED_ASSETS } from '../utils/blacklist'
-import { SUPPORTED_CHAIN_IDS } from '../utils/constants'
 
 export const filterAssetIdsBySellable = (assets: Asset[]): AssetId[] => {
   return assets
