+++ packages/swapper/src/swappers/CowSwapper/filterBuyAssetsBySellAssetId/filterBuyAssetsBySellAssetId.ts
@@ -3,7 +3,7 @@ import type { AssetId } from '@shapeshiftoss/caip'
 import type { BuyAssetBySellIdInput } from '../../../types'
 import { isNativeEvmAsset } from '../../utils/helpers/helpers'
 import { COWSWAP_UNSUPPORTED_ASSETS } from '../utils/blacklist'
-import { SUPPORTED_CHAIN_IDS } from '../utils/constants'
+import { SUPPORTED_CHAIN_IDS } from '../../../cowswap-utils/constants'
 
 export const filterBuyAssetsBySellAssetId = ({
   assets,
