+++ packages/swapper/src/swappers/CowSwapper/getCowSwapTradeQuote/getCowSwapTradeQuote.ts
@@ -13,6 +13,11 @@ import {
   getCowNetwork,
   getFullAppData,
 } from '../../../cowswap-utils'
+import {
+  COW_SWAP_NATIVE_ASSET_MARKER_ADDRESS,
+  COW_SWAP_VAULT_RELAYER_ADDRESS,
+  SUPPORTED_CHAIN_IDS,
+} from '../../../cowswap-utils/constants'
 import type {
   GetEvmTradeQuoteInputBase,
   SwapErrorRight,
@@ -22,11 +27,6 @@ import type {
 import { SwapperName, TradeQuoteError } from '../../../types'
 import { createTradeAmountTooSmallErr, makeSwapErrorRight } from '../../../utils'
 import { isNativeEvmAsset } from '../../utils/helpers/helpers'
-import {
-  COW_SWAP_NATIVE_ASSET_MARKER_ADDRESS,
-  COW_SWAP_VAULT_RELAYER_ADDRESS,
-  SUPPORTED_CHAIN_IDS,
-} from '../utils/constants'
 import { cowService } from '../utils/cowService'
 import {
   assertValidTrade,
