+++ packages/swapper/src/swappers/CowSwapper/utils/helpers/helpers.ts
@@ -15,21 +15,6 @@ import type { SwapErrorRight } from '../../../../types'
 import { TradeQuoteError } from '../../../../types'
 import { makeSwapErrorRight } from '../../../../utils'
 
-export const ORDER_TYPE_FIELDS = [
-  { name: 'sellToken', type: 'address' },
-  { name: 'buyToken', type: 'address' },
-  { name: 'receiver', type: 'address' },
-  { name: 'sellAmount', type: 'uint256' },
-  { name: 'buyAmount', type: 'uint256' },
-  { name: 'validTo', type: 'uint32' },
-  { name: 'appData', type: 'bytes32' },
-  { name: 'feeAmount', type: 'uint256' },
-  { name: 'kind', type: 'string' },
-  { name: 'partiallyFillable', type: 'bool' },
-  { name: 'sellTokenBalance', type: 'string' },
-  { name: 'buyTokenBalance', type: 'string' },
-]
-
 export type CowSwapQuoteApiInputBase = {
   appData: string
   appDataHash?: string
